@@ -6,12 +6,12 @@ use Laralabs\GetAddress\GetAddressServiceProvider;
 
 abstract class TestCase extends \Orchestra\Testbench\TestCase
 {
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
     }
 
-    protected function getPackageProviders($app)
+    protected function getPackageProviders($app): array
     {
         return [
             GetAddressServiceProvider::class,
